Load xhr file from assets
